# Conflicts:
#	modules/apps/web-experience/staging/staging-configuration-web/src/main/resources/META-INF/resources/view.jsp
#	modules/apps/web-experience/staging/staging-taglib/src/main/resources/META-INF/resources/liferay-staging.tld

+++ modules/apps/web-experience/fragment/fragment-service/src/main/java/com/liferay/fragment/service/impl/FragmentEntryLocalServiceImpl.java
@@ -18,6 +18,7 @@ import com.liferay.fragment.exception.DuplicateFragmentEntryException;
 import com.liferay.fragment.exception.FragmentEntryNameException;
 import com.liferay.fragment.exception.RequiredFragmentEntryException;
 import com.liferay.fragment.model.FragmentEntry;
+import com.liferay.fragment.processor.FragmentEntryProcessorRegistry;
 import com.liferay.fragment.service.base.FragmentEntryLocalServiceBaseImpl;
 import com.liferay.html.preview.model.HtmlPreviewEntry;
 import com.liferay.html.preview.service.HtmlPreviewEntryLocalService;
@@ -28,6 +29,7 @@ import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.OrderByComparator;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.spring.extender.service.ServiceReference;
 
 import java.util.Date;
@@ -43,14 +45,59 @@ public class FragmentEntryLocalServiceImpl
 	@Override
 	public FragmentEntry addFragmentEntry(
 			long userId, long groupId, long fragmentCollectionId, String name,
-			String css, String html, String js, ServiceContext serviceContext)
+			int status, ServiceContext serviceContext)
 		throws PortalException {
 
 		// Fragment entry
 
 		User user = userLocalService.getUser(userId);
 
-		validate(groupId, name);
+		validate(groupId, fragmentCollectionId, 0, name);
+
+		long fragmentEntryId = counterLocalService.increment();
+
+		FragmentEntry fragmentEntry = fragmentEntryPersistence.create(
+			fragmentEntryId);
+
+		fragmentEntry.setGroupId(groupId);
+		fragmentEntry.setCompanyId(user.getCompanyId());
+		fragmentEntry.setUserId(user.getUserId());
+		fragmentEntry.setUserName(user.getFullName());
+		fragmentEntry.setCreateDate(serviceContext.getCreateDate(new Date()));
+		fragmentEntry.setModifiedDate(
+			serviceContext.getModifiedDate(new Date()));
+		fragmentEntry.setFragmentCollectionId(fragmentCollectionId);
+		fragmentEntry.setName(name);
+		fragmentEntry.setStatus(status);
+		fragmentEntry.setStatusByUserId(userId);
+		fragmentEntry.setStatusByUserName(user.getFullName());
+		fragmentEntry.setStatusDate(new Date());
+
+		fragmentEntryPersistence.update(fragmentEntry);
+
+		// Resources
+
+		resourceLocalService.addModelResources(fragmentEntry, serviceContext);
+
+		return fragmentEntry;
+	}
+
+	@Override
+	public FragmentEntry addFragmentEntry(
+			long userId, long groupId, long fragmentCollectionId, String name,
+			String css, String html, String js, int status,
+			ServiceContext serviceContext)
+		throws PortalException {
+
+		// Fragment entry
+
+		User user = userLocalService.getUser(userId);
+
+		validate(groupId, fragmentCollectionId, 0, name);
+
+		if (WorkflowConstants.STATUS_APPROVED == status) {
+			validateContent(html);
+		}
 
 		long fragmentEntryId = counterLocalService.increment();
 
@@ -69,8 +116,10 @@ public class FragmentEntryLocalServiceImpl
 		fragmentEntry.setCss(css);
 		fragmentEntry.setHtml(html);
 		fragmentEntry.setJs(js);
-
-		// HTML preview
+		fragmentEntry.setStatus(status);
+		fragmentEntry.setStatusByUserId(userId);
+		fragmentEntry.setStatusByUserName(user.getFullName());
+		fragmentEntry.setStatusDate(new Date());
 
 		HtmlPreviewEntry htmlPreviewEntry = _updateHtmlPreviewEntry(
 			fragmentEntry, serviceContext);
@@ -112,8 +161,10 @@ public class FragmentEntryLocalServiceImpl
 
 		// HTML preview
 
-		_htmlPreviewEntryLocalService.deleteHtmlPreviewEntry(
-			fragmentEntry.getHtmlPreviewEntryId());
+		if (fragmentEntry.getHtmlPreviewEntryId() > 0) {
+			_htmlPreviewEntryLocalService.deleteHtmlPreviewEntry(
+				fragmentEntry.getHtmlPreviewEntryId());
+		}
 
 		return fragmentEntry;
 	}
@@ -128,14 +179,22 @@ public class FragmentEntryLocalServiceImpl
 	}
 
 	@Override
-	public List<FragmentEntry> fetchFragmentEntries(long fragmentCollectionId) {
+	public FragmentEntry fetchFragmentEntry(long fragmentEntryId) {
+		return fragmentEntryPersistence.fetchByPrimaryKey(fragmentEntryId);
+	}
+
+	@Override
+	public List<FragmentEntry> getFragmentEntries(long fragmentCollectionId) {
 		return fragmentEntryPersistence.findByFragmentCollectionId(
 			fragmentCollectionId);
 	}
 
 	@Override
-	public FragmentEntry fetchFragmentEntry(long fragmentEntryId) {
-		return fragmentEntryPersistence.fetchByPrimaryKey(fragmentEntryId);
+	public List<FragmentEntry> getFragmentEntries(
+		long fragmentCollectionId, int status) {
+
+		return fragmentEntryPersistence.findByFCI_S(
+			fragmentCollectionId, status);
 	}
 
 	@Override
@@ -172,6 +231,52 @@ public class FragmentEntryLocalServiceImpl
 	}
 
 	@Override
+	public int getFragmentEntriesCount(long fragmentCollectionId) {
+		return fragmentEntryPersistence.countByFragmentCollectionId(
+			fragmentCollectionId);
+	}
+
+	@Override
+	public FragmentEntry updateFragmentEntry(
+			long userId, long fragmentEntryId, String name, String css,
+			String html, String js, int status, ServiceContext serviceContext)
+		throws PortalException {
+
+		FragmentEntry fragmentEntry = fragmentEntryPersistence.findByPrimaryKey(
+			fragmentEntryId);
+
+		validate(
+			fragmentEntry.getGroupId(), fragmentEntry.getFragmentCollectionId(),
+			fragmentEntryId, name);
+
+		if (WorkflowConstants.STATUS_APPROVED == status) {
+			validateContent(html);
+		}
+
+		User user = userLocalService.getUser(userId);
+
+		fragmentEntry.setModifiedDate(new Date());
+		fragmentEntry.setName(name);
+		fragmentEntry.setCss(css);
+		fragmentEntry.setHtml(html);
+		fragmentEntry.setJs(js);
+		fragmentEntry.setStatus(status);
+		fragmentEntry.setStatusByUserId(userId);
+		fragmentEntry.setStatusByUserName(user.getFullName());
+		fragmentEntry.setStatusDate(new Date());
+
+		HtmlPreviewEntry htmlPreviewEntry = _updateHtmlPreviewEntry(
+			fragmentEntry, serviceContext);
+
+		fragmentEntry.setHtmlPreviewEntryId(
+			htmlPreviewEntry.getHtmlPreviewEntryId());
+
+		fragmentEntryPersistence.update(fragmentEntry);
+
+		return fragmentEntry;
+	}
+
+	@Override
 	public FragmentEntry updateFragmentEntry(long fragmentEntryId, String name)
 		throws PortalException {
 
@@ -182,51 +287,37 @@ public class FragmentEntryLocalServiceImpl
 			return fragmentEntry;
 		}
 
-		validate(fragmentEntry.getGroupId(), name);
+		validate(
+			fragmentEntry.getGroupId(), fragmentEntry.getFragmentCollectionId(),
+			fragmentEntryId, name);
 
 		fragmentEntry.setName(name);
 
 		return fragmentEntryPersistence.update(fragmentEntry);
 	}
 
-	@Override
-	public FragmentEntry updateFragmentEntry(
-			long fragmentEntryId, String name, String css, String html,
-			String js, ServiceContext serviceContext)
+	protected void validate(
+			long groupId, long fragmentCollectionId, long fragmentEntryId,
+			String name)
 		throws PortalException {
 
-		FragmentEntry fragmentEntry = fragmentEntryPersistence.findByPrimaryKey(
-			fragmentEntryId);
-
-		if (!Objects.equals(fragmentEntry.getName(), name)) {
-			validate(fragmentEntry.getGroupId(), name);
-		}
-
-		fragmentEntry.setModifiedDate(new Date());
-		fragmentEntry.setName(name);
-		fragmentEntry.setCss(css);
-		fragmentEntry.setHtml(html);
-		fragmentEntry.setJs(js);
-
-		_updateHtmlPreviewEntry(fragmentEntry, serviceContext);
-
-		fragmentEntryPersistence.update(fragmentEntry);
-
-		return fragmentEntry;
-	}
-
-	protected void validate(long groupId, String name) throws PortalException {
 		if (Validator.isNull(name)) {
 			throw new FragmentEntryNameException(
 				"Name must not be null for group " + groupId);
 		}
 
-		FragmentEntry fragmentEntry = fragmentEntryPersistence.fetchByG_N(
-			groupId, name);
+		FragmentEntry fragmentEntry = fragmentEntryPersistence.fetchByG_FCI_N(
+			groupId, fragmentCollectionId, name);
 
-		if (fragmentEntry != null) {
+		if ((fragmentEntry != null) &&
+			(fragmentEntry.getFragmentEntryId() != fragmentEntryId)) {
+
 			throw new DuplicateFragmentEntryException(name);
 		}
+	}
+
+	protected void validateContent(String html) throws PortalException {
+		_fragmentEntryProcessorRegistry.validateFragmentEntryHTML(html);
 	}
 
 	private HtmlPreviewEntry _updateHtmlPreviewEntry(
@@ -250,6 +341,9 @@ public class FragmentEntryLocalServiceImpl
 			fragmentEntry.getFragmentEntryId(), fragmentEntry.getContent(),
 			ContentTypes.IMAGE_PNG, serviceContext);
 	}
+
+	@ServiceReference(type = FragmentEntryProcessorRegistry.class)
+	private FragmentEntryProcessorRegistry _fragmentEntryProcessorRegistry;
 
 	@ServiceReference(type = HtmlPreviewEntryLocalService.class)
 	private HtmlPreviewEntryLocalService _htmlPreviewEntryLocalService;
